Updated To Include Model Grid
Updated program to include ability to accept model grid aquifer properties. If model_grid passed as an argument then instead of taking the properties assigned at the well a line will be drawn between the well and the considered reach, and the properties at the grid along that line averaged and used.