Fixed issue with Thiessen polygon apportionment
Issue would arise with wells far from streams, where the default clipping of thiessen polygons would make apportioned fractions in some instances not sum to 1. Added envelope argument on line 1541 to address this.

Additionally added code to handle if there is only one reach within the zone of a wells influence. Previously the code would run in unpredictable ways, as a thiessen polygon cannot be made out of one entity. Now the code skips generating thiessen polygons in the case of one entity, and simply gives it a fraction of 1.